:art: staff class create and refactor classes
